+++ cdm_reader_mapper/common/replace.py
@@ -19,7 +19,6 @@ Replacement arguments:
 
 from __future__ import annotations
 
-
 import pandas as pd
 
 from . import logging_hdlr