+++ cdm_reader_mapper/mdf_reader/writer.py
@@ -6,28 +6,66 @@ import json
 import logging
 from io import StringIO as StringIO
 from pathlib import Path
+from typing import Any, get_args
 
 import pandas as pd
+from pandas.io.parsers import TextFileReader
 
 from .utils.utilities import join, update_column_names, update_dtypes
 
 from ..common import get_filename
 from ..common.pandas_TextParser_hdlr import make_copy
 
+from ..properties import SupportedFileTypes
+
+WRITERS = {
+    "csv": "to_csv",
+    "parquet": "to_parquet",
+    "feather": "to_feather",
+}
+
+
+def _normalize_data_chunks(
+    data: pd.DataFrame | TextFileReader | None,
+) -> list | TextFileReader:
+    """Helper function to normalize data chunks."""
+    if data is None:
+        data = pd.DataFrame()
+    if isinstance(data, pd.DataFrame):
+        return [data]
+    if isinstance(data, TextFileReader):
+        return make_copy(data)
+    raise TypeError(f"Unsupported data type found: {type(data)}.")
+
+
+def _write_data(
+    data_df: pd.DataFrame,
+    mask_df: pd.DataFrame,
+    data_fn: str,
+    mask_fn: str,
+    writer: str,
+    write_kwargs: dict[str, Any],
+) -> None:
+    """Helper function to write data on disk."""
+    getattr(data_df, writer)(data_fn, **write_kwargs)
+    if not mask_df.empty:
+        getattr(mask_df, writer)(mask_fn, **write_kwargs)
+
 
 def write_data(
-    data,
-    mask=None,
-    dtypes: dict | None = None,
+    data: pd.DataFrame | TextFileReader,
+    mask: pd.DataFrame | TextFileReader | None = None,
+    data_format: SupportedFileTypes = "csv",
+    dtypes: pd.Series | dict | None = None,
     parse_dates: list | bool = False,
-    encoding="utf-8",
-    out_dir=".",
-    prefix=None,
-    suffix=None,
-    extension="csv",
-    filename=None,
-    col_subset=None,
-    delimiter=",",
+    encoding: str = "utf-8",
+    out_dir: str = ".",
+    prefix: str | None = None,
+    suffix: str | None = None,
+    extension: str = None,
+    filename: str | dict | None = None,
+    col_subset: str | list | tuple | None = None,
+    delimiter: str = ",",
     **kwargs,
 ) -> None:
     """Write pandas.DataFrame to MDF file on file system.
@@ -36,31 +74,32 @@ def write_data(
     ----------
     data: pandas.DataFrame
         pandas.DataFrame to export.
-    mask: pandas.DataFrame
+    mask: pandas.DataFrame, optional
         validation mask to export.
-    dtypes: dict
+    data_format: {"csv", "parquet", "feather"}, default: "csv"
+        Format of output data file(s).
+    dtypes: dict, optional
         Dictionary of data types on ``data``.
         Dump ``dtypes`` and ``parse_dates`` to json information file.
-    parse_dates:
+    parse_dates: list | bool, default: False
         Information of how to parse dates in :py:attr:`data`.
         Dump ``dtypes`` and ``parse_dates`` to json information file.
         For more information see :py:func:`pandas.read_csv`.
-    encoding: str
+    encoding: str, default: "utf-8"
         A string representing the encoding to use in the output file, defaults to utf-8.
-    out_dir: str
+    out_dir: str, default: "."
         Path to the output directory.
-        Default: current directory
     prefix: str, optional
         Prefix of file name structure: ``<prefix>-data-*<suffix>.<extension>``.
     suffix: str, optional
         Suffix of file name structure: ``<prefix>-data-*<suffix>.<extension>``.
-    extension: str
+    extension: str, optional
         Extension of file name structure: ``<prefix>-data-*<suffix>.<extension>``.
-        Default: psv
+        By default, extension depends on `data_format`.
     filename: str or dict, optional
         Name of the output file name(s).
         List one filename for both ``data`` and ``mask`` ({"data":<filenameD>, "mask":<filenameM>}).
-        Default: Automatically create file name from table name, ``prefix`` and ``suffix``.
+        By default, automatically create file name from table name, ``prefix`` and ``suffix``.
     col_subset: str, tuple or list, optional
         Specify the section or sections of the file to write.
 
@@ -71,9 +110,8 @@ def write_data(
           e.g. list type object col_subset = [columns]
 
         Column labels could be both string or tuple.
-    delimiter: str
+    delimiter: str, default: ","
         Character or regex pattern to treat as the delimiter while reading with df.to_csv.
-        Default: ","
 
     See Also
     --------
@@ -88,24 +126,27 @@ def write_data(
     ----
     Use this function after reading MDF data.
     """
-    dtypes = dtypes or {}
+    supported_file_types = get_args(SupportedFileTypes)
+    if data_format not in supported_file_types:
+        raise ValueError(
+            f"data_format must be one of {supported_file_types}, not {data_format}."
+        )
+
+    extension = extension or data_format
+
+    if not isinstance(dtypes, (dict, pd.Series)):
+        dtypes = {}
+
     if isinstance(parse_dates, bool):
         parse_dates = []
 
-    if not isinstance(data, pd.io.parsers.TextFileReader):
-        data_list = [data]
-    else:
-        data_list = make_copy(data)
+    data_list = _normalize_data_chunks(data)
+    mask_list = _normalize_data_chunks(mask)
 
-    if mask is None:
-        mask = pd.DataFrame()
-
-    if not isinstance(mask, pd.io.parsers.TextFileReader):
-        mask_list = [mask]
-    else:
-        mask_list = make_copy(mask)
-
-    info = {"dtypes": dtypes.copy(), "parse_dates": [join(p) for p in parse_dates]}
+    info = {
+        "dtypes": {k: str(v) for k, v in dtypes.items()},
+        "parse_dates": [join(p) for p in parse_dates],
+    }
 
     logging.info(f"WRITING DATA TO FILES IN: {out_dir}")
     out_dir = Path(out_dir)
@@ -142,18 +183,25 @@ def write_data(
             info["parse_dates"] = [p for p in info["parse_dates"] if p in header]
             info["encoding"] = encoding
 
-        csv_kwargs = dict(
-            header=header,
-            mode=mode,
-            index=False,
-            sep=delimiter,
-            encoding=encoding,
-            **kwargs,
-        )
+        write_kwargs = {}
+        if data_format == "csv":
+            write_kwargs = dict(
+                header=header,
+                mode=mode,
+                index=False,
+                sep=delimiter,
+                encoding=encoding,
+                **kwargs,
+            )
 
-        data_df.to_csv(filename_data, **csv_kwargs)
-        if not mask_df.empty:
-            mask_df.to_csv(filename_mask, **csv_kwargs)
+        _write_data(
+            data_df=data_df,
+            mask_df=mask_df,
+            data_fn=filename_data,
+            mask_fn=filename_mask,
+            writer=WRITERS[data_format],
+            write_kwargs=write_kwargs,
+        )
 
     with open(filename_info, "w") as fileObj:
         json.dump(info, fileObj, indent=4)